Fix implicit conversion
error: implicit conversion loses integer precision: 'long' to 'int'.
error: implicit conversion loses integer precision: 'unsigned long' to
'int'.

+++ src/cmp.c
@@ -6,7 +6,7 @@
 /*   By: sblauens <sblauens@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/29 13:53:52 by sblauens          #+#    #+#             */
-/*   Updated: 2018/07/04 04:39:03 by sblauens         ###   ########.fr       */
+/*   Updated: 2018/07/04 06:25:32 by sblauens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@ static int						cmp_files_time(t_timespec *t1, t_timespec *t2)
 {
 	int		time_cmp;
 
-	time_cmp = t2->tv_sec - t1->tv_sec;
+	time_cmp = (int)(t2->tv_sec - t1->tv_sec);
 	if (!time_cmp)
-		time_cmp = t2->tv_nsec - t1->tv_nsec;
+		time_cmp = (int)(t2->tv_nsec - t1->tv_nsec);
 	return (time_cmp);
 }
